pyproject.toml
[project]
name = "requirements-advisor"
version = "0.1.0"
description = "MCP server providing expert guidance on requirements management best practices"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["mcp", "requirements-management", "rag", "llm", "jama"]

dependencies = [
    "fastmcp>=0.4.0",
    "httpx>=0.27.0",
    "voyageai>=0.3.0",
    "chromadb>=0.5.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.2.0",
    "uvicorn>=0.29.0",
    "rich>=13.7.0",
    "typer>=0.12.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "respx>=0.21.0",
]
ingestion = [
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "tenacity>=8.2.0",
    "Pillow>=10.0.0",
]

[project.scripts]
requirements-advisor = "requirements_advisor.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/requirements_advisor"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
